Initial commit
Verilog code to run fp_mul module on a Digilent Basys 3 FPGA board.
The fp_mul module is configured to support the IEEE 754 16-bit
binary format. Operands can be input from either the 16 user switches
or from the product generated by fp_mul. See the README.md file for
information about the video which describes the use of the code.